Segment padding error message clarity.
It wasn't clear what "crypto padding" meant and if it was referring to the token or the key. This is a small change that indicates that the error is in the crypto portion of the token.

I ran into this issue while trying to use the RS256 algorithm where upon I was trying to figure out what that error meant and most of the results that came back all referenced issues with people using non-pem formatted keys. This led me down a false path where in I thought there was something wrong with my key as "crypto" (as well as payload) aren't indicators of where the error is occurring (in the token.)