Update typing syntax and usage for Python 3.6+
Now that Python 2 is not supported, can move away from type comments to
type annotation 🎉.

The typing module is always available, so remove the guards.

Specify the supported Python in the mypy configuration.

Move other mypy configurations to one place. This way, whether tox is
used or not, the same mypy errors appear.

Distribute and install PEP-561 compliant py.typed file. When PyJWT is a
imported as a library, this tells mypy to use the provided type
annotations rather than going through typeshed. This way, the types are
always up to date when running mypy.

Remove outdated ignores since dropping Python 2.